Fix: Circular dependency
